+++ src/models/blocks/document-body-image.ts
@@ -1,4 +1,5 @@
 import { DocumentBodyBlockAlignType } from './document-body-block.js';
+import { DocumentElementLength } from './document-element-length.js';
 
 export interface DocumentBodyImageBlock {
   type: 'Image';
@@ -16,4 +17,5 @@ export interface DocumentBodyImageProperties {
   align?: DocumentBodyBlockAlignType;
   indentation?: number;
   width?: number;
+  widthWithUnit?: DocumentElementLength;
 }