+++ src/models/blocks/document-element-length.ts
@@ -1,0 +1,10 @@
+export enum DocumentElementLengthUnit {
+  Percentage = 'Percentage',
+  Em = 'Em',
+  Px = 'Px',
+}
+
+export interface DocumentElementLength {
+  value: number;
+  unit: DocumentElementLengthUnit;
+}