+++ src/models/html/length.ts
@@ -1,6 +1,0 @@
-import { DocumentBodyLengthUnit } from '../blocks/document-body-table.js';
-
-export interface Length {
-  length: number;
-  unit?: DocumentBodyLengthUnit;
-}